No round robin approach within a single timestamp
This change has been a long time coming. We observed during the BGP hackathon last year that it was faster to have many processes each with only a single collector than it was to have a single process with many collectors. One of the main reasons for this is that we were (pointlessly) round-robin'ing between files with every record. This change continues to pop records from a single file until the timestamp changes.