Make code style consistent (#35)
As noted in Issue #24, we were in need of a consistent code style throughout the project.

This adds a [clang-format](http://clang.llvm.org/docs/ClangFormat.html) configuration file (`.clang-format`) and a `format` make target to automatically reformat all code. It also re-formats all existing code to match this style.

The style that I have adopted is very similar to the [LLVM style](http://llvm.org/docs/CodingStandards.html), but see the `.clang-format` file added by this PR for definitive information.

I'm also happy to hear arguments for changing a specific aspect of the style.